Move test doubles into their own files
