+++ EurofurenceTests/Presenter Tests/Message Detail/MessageDetailPresenterTests.swift
@@ -3,136 +3,6 @@ import EurofurenceModel
 import EurofurenceModelTestDoubles
 import XCTest
 
-class SuccessfulPrivateMessagesService: PrivateMessagesService {
-    
-    private let expected: MessageIdentifier
-    private let response: Result<Message, PrivateMessageError>
-    
-    init(successfulForMessage expected: MessageIdentifier, providingMessage returnValue: Message) {
-        self.expected = expected
-        self.response = .success(returnValue)
-    }
-    
-    func add(_ observer: PrivateMessagesObserver) {
-        
-    }
-    
-    func refreshMessages() {
-        
-    }
-    
-    func fetchMessage(identifiedBy identifier: MessageIdentifier, completionHandler: @escaping (Result<Message, PrivateMessageError>) -> Void) {
-        if identifier == expected {
-            completionHandler(response)
-        }
-    }
-    
-}
-
-class FailingPrivateMessagesService: PrivateMessagesService {
-    
-    private let expected: MessageIdentifier
-    private let response: Result<Message, PrivateMessageError>
-    
-    init(unsuccessfulForMessage expected: MessageIdentifier, providingError returnValue: PrivateMessageError) {
-        self.expected = expected
-        self.response = .failure(returnValue)
-    }
-    
-    func add(_ observer: PrivateMessagesObserver) {
-        
-    }
-    
-    func refreshMessages() {
-        
-    }
-    
-    func fetchMessage(identifiedBy identifier: MessageIdentifier, completionHandler: @escaping (Result<Message, PrivateMessageError>) -> Void) {
-        if identifier == expected {
-            completionHandler(response)
-        }
-    }
-    
-}
-
-class ControllablePrivateMessagesService: PrivateMessagesService {
-    
-    func add(_ observer: PrivateMessagesObserver) {
-        
-    }
-    
-    func refreshMessages() {
-        
-    }
-    
-    private var currentCompletionHandler: ((Result<Message, PrivateMessageError>) -> Void)?
-    func fetchMessage(identifiedBy identifier: MessageIdentifier, completionHandler: @escaping (Result<Message, PrivateMessageError>) -> Void) {
-        currentCompletionHandler = completionHandler
-    }
-    
-    func succeedNow(message: Message) {
-        currentCompletionHandler?(.success(message))
-        currentCompletionHandler = nil
-    }
-    
-    func failNow(error: PrivateMessageError) {
-        currentCompletionHandler?(.failure(error))
-        currentCompletionHandler = nil
-    }
-    
-}
-
-struct StubMessageDetailSceneFactory: MessageDetailSceneFactory {
-
-    let scene = CapturingMessageDetailScene()
-    func makeMessageDetailScene() -> UIViewController & MessageDetailScene {
-        return scene
-    }
-
-}
-
-class CapturingMessageDetailScene: UIViewController, MessageDetailScene {
-
-    var delegate: MessageDetailSceneDelegate?
-    
-    private(set) var loadingIndicatorVisibility: VisibilityState = .unset
-    private(set) var viewModel: MessageDetailViewModel?
-    private(set) var errorViewModel: MessageDetailErrorViewModel?
-    
-    func showLoadingIndicator() {
-        loadingIndicatorVisibility = .visible
-    }
-    
-    func hideLoadingIndicator() {
-        loadingIndicatorVisibility = .hidden
-    }
-    
-    func showMessage(viewModel: MessageDetailViewModel) {
-        self.viewModel = viewModel
-    }
-    
-    func showError(viewModel: MessageDetailErrorViewModel) {
-        errorViewModel = viewModel
-    }
-
-    private(set) var capturedMessageDetailTitle: String?
-    func setMessageDetailTitle(_ title: String) {
-        capturedMessageDetailTitle = title
-    }
-
-    private(set) var numberOfMessageComponentsAdded = 0
-    private(set) var capturedMessageBinder: MessageComponentBinder?
-    func addMessageComponent(with binder: MessageComponentBinder) {
-        numberOfMessageComponentsAdded += 1
-        capturedMessageBinder = binder
-    }
-    
-    func simulateSceneReady() {
-        delegate?.messageDetailSceneDidLoad()
-    }
-    
-}
-
 class MessageDetailPresenter2Tests: XCTestCase {
     
     func testLoadingMessage() {
