+++ EurofurenceTests/Presenter Tests/Message Detail/Test Doubles/StubMessageDetailSceneFactory.swift
@@ -1,0 +1,11 @@
+import Eurofurence
+import UIKit
+
+struct StubMessageDetailSceneFactory: MessageDetailSceneFactory {
+
+    let scene = CapturingMessageDetailScene()
+    func makeMessageDetailScene() -> UIViewController & MessageDetailScene {
+        return scene
+    }
+
+}
